Transport panel is available but not styled correctly
