+++ yukon/html/monitor/monitor.html
@@ -10,6 +10,7 @@
     <script src="../api.js"></script>
     <script src="../autosize.js"></script>
     <script src="../tween.cjs.js"></script>
+    <script src="../tween.js"></script>
     <script src="../golden-layout/golden-layout.min.js"></script>
     <script src="messages/messages.js"></script>
     <script src="../js-yaml.js"></script>
@@ -32,7 +33,8 @@
     <link rel="stylesheet" href="../cool-custom-context-menu/ContextMenu.css" />
     <link rel="stylesheet" href="../golden-layout/goldenlayout-base.css">
     <link rel="stylesheet" href="../golden-layout/goldenlayout-light-theme.css">
-    
+    <link rel="stylesheet" href="style/transport.css">
+
     <title>Yukon — Monitor</title>
 </head>
 
@@ -81,10 +83,10 @@
             <textarea id="textOut" spellcheck="false"></textarea>
         </div>
         <div class="c-tab">
-            
+
         </div>
         <div class="c-tab registers-tab">
-            
+
         </div>
     </div>
