+++ yukon/html/monitor/style/transport.css
@@ -9,12 +9,18 @@ body {
 }
 
 .noselect * {
-  -webkit-touch-callout: none; /* iOS Safari */
-    -webkit-user-select: none; /* Safari */
-     -khtml-user-select: none; /* Konqueror HTML */
-       -moz-user-select: none; /* Old versions of Firefox */
-        -ms-user-select: none; /* Internet Explorer/Edge */
-            user-select: none; /* Non-prefixed version, currently
+    -webkit-touch-callout: none;
+    /* iOS Safari */
+    -webkit-user-select: none;
+    /* Safari */
+    -khtml-user-select: none;
+    /* Konqueror HTML */
+    -moz-user-select: none;
+    /* Old versions of Firefox */
+    -ms-user-select: none;
+    /* Internet Explorer/Edge */
+    user-select: none;
+    /* Non-prefixed version, currently
                                   supported by Chrome, Edge, Opera and Firefox */
 }
 
@@ -116,7 +122,7 @@ input[type=text] {
     display: inline-block !important;
 }
 
-form#transport-selection-form > #maybe-tabs > div.tab-row {
+form#transport-selection-form>#maybe-tabs>div.tab-row {
     overflow: hidden;
     position: relative;
     /* Making it an anchor point for tab-slider */
